+++ src/main/java/com/silvaguilherme/estudoshorasepc/entities/Setor.java
@@ -18,7 +18,7 @@ public class Setor implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String name;
 	
 	@JsonIgnore
@@ -28,17 +28,17 @@ public class Setor implements Serializable {
 	public Setor() {
 	}
 
-	public Setor(Long id, String name) {
+	public Setor(Integer id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
